♻️ [RUM-232] move deflate worker implementation to startDeflateWorker
The old `deflateWorker.js` module was there mainly to contain pako and
the worker implementation. Now that it's been moved, there is little
benefit to keep the old module. Putting the code to
`startDeflateWorker.ts` removes the confusion between the two modules.
